feat: add support for required keyword
BREAKING CHANGES:
Prior to this change the nodejs api returned added and removed differences detected as as an array of DiffResultDifference objects. Now the added and removed differences are returned as two seperate JsonSchema objects. In addition all the property names of the DiffResult object have been renamed. To migrate, update code that consumes the DiffResult object returned by the API to the new property names and types.